% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvd.R
\name{csvd}
\alias{csvd}
\title{Constrained SVD of a matrix (wrapper of c++ functions).}
\usage{
csvd(
  X,
  R = 2,
  au = rep(1.4, R),
  av = rep(1.4, R),
  itermax.pi = 1000,
  itermax.pocs = 1000,
  eps.pi = 1e-16,
  eps.pocs = 1e-16,
  init = "svd"
)
}
\arguments{
\item{X}{a (data) matrix;}

\item{R}{the desired rank of the singular decomposition;}

\item{au}{The radiuses (radii?) (>0) of the
$L_1$ ball for each left vector}

\item{av}{The radiuses (radii)?
(>0) of the $L_1$ balls for each right vector}

\item{itermax.pi}{The maximum number of power iterations}

\item{itermax.pocs}{The maximum number of projection iterations}

\item{eps.pi}{The power iteration precision}

\item{eps.pocs}{The projection precision}

\item{init}{How to initialize the algorithm}
}
\value{
Pseudo-singular vectors and values
}
\description{
Constrained SVD of a matrix (wrapper of c++ functions).
}
\examples{
X <- matrix(rnorm(20), 5, 4)
csvd(X)
}
\author{
Vincent Guillemot
}
