% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm}
\alias{norm}
\alias{norm1}
\alias{norm2}
\alias{normgroup}
\alias{normalize}
\title{Compute the l1 norm of a vector.}
\usage{
norm1(u)

norm2(u)

normgroup(u, g)

normalize(u)
}
\arguments{
\item{u}{A vector of numerics}

\item{g}{A group factor}
}
\value{
The l1 norm of \eqn{u}: \eqn{\sum_i |u_i|}.

The l2 norm of \eqn{u}: \eqn{\sqrt{\sum_i u_i^2}}.

The \eqn{l_{1,2}} group norm of \eqn{u}: \eqn{\sum_g ||u_g||_2}.

The normalized version of \eqn{u}.
}
\description{
Compute the l1 norm of a vector.

Compute the l2 norm of a vector.

Compute the l1,2 group norm of a vector.

Normalize a vector.
}
\examples{
norm1(c(0, 0, -0.5, 0.5))
norm2(c(0, 0, -0.5, 0.5))
normgroup(c(0, 0, -0.5, 0.5), gl(2, 2))
normalize(c(0, 0, -0.5, 0.5))
}
