% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm1.R
\name{norm1}
\alias{norm1}
\title{Compute the l1 norm of a vector.}
\usage{
norm1(u)
}
\arguments{
\item{u}{A vector of numerics}
}
\value{
The \eqn{l_1} norm of \eqn{u}: \eqn{\sum_i |u_i|}.
}
\description{
Compute the l1 norm of a vector.
}
\examples{
norm1( c(0, 0, -0.5, 0.5) )
}

