% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm2.R
\name{norm2}
\alias{norm2}
\title{Compute the l2 norm of a vector.}
\usage{
norm2(u)
}
\arguments{
\item{u}{A vector of numerics}
}
\value{
The l2 norm of \eqn{u}: \eqn{\sqrt{\sum_i u_i^2}}.
}
\description{
Compute the l2 norm of a vector.
}
\examples{
norm2( c(0, 0, -0.5, 0.5) )
}

