% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normgroup.R
\name{normgroup}
\alias{normgroup}
\title{Compute the l1,2 group norm of a vector.}
\usage{
normgroup(u, g)
}
\arguments{
\item{u}{A vector of numerics}

\item{g}{A group factor}
}
\value{
The \eqn{l_{1,2}} group norm of \eqn{u}: \eqn{\sum_g ||u_g||_2}.
}
\description{
Compute the l1,2 group norm of a vector.
}
\examples{
normgroup( c(0, 0, -0.5, 0.5), gl(2, 2) )
}
