% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\alias{powit1}
\alias{powit2}
\title{Compute one constrained power iteration.}
\usage{
powit1(X, U0, V0, au, av, eps.pi, itermax.pi)

powit2(
  X,
  U0,
  V0,
  Uorth,
  Vorth,
  au,
  av,
  eps.pi,
  eps.pocs,
  itermax.pi,
  itermax.pocs
)
}
\arguments{
\item{X}{A matrix}

\item{U0}{Initial left singular vector}

\item{V0}{Initial right singular vector}

\item{au}{Left-vector radius}

\item{av}{Right-vector radius}

\item{eps.pi}{Power iteration precision}

\item{itermax.pi}{Maximum number of power iterations}

\item{Uorth}{Existing left vectors}

\item{Vorth}{Existing right vectors}

\item{eps.pocs}{Projection precision}

\item{itermax.pocs}{Maximum number of projection iterations}
}
\value{
A list containing the updated vectors and iteration count.
}
\description{
Compute one constrained power iteration.

Compute one orthogonally constrained power iteration.
}
