% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj12.R
\name{proj12}
\alias{proj12}
\title{Compute the projection of a vector onto the intersection of the l1 ball
and the l2 ball.}
\usage{
proj12(x, a = 1)
}
\arguments{
\item{x}{A vector of numerics}

\item{a}{The radius (>0) of the l1 ball}
}
\value{
The "projection" of \eqn{x} onto \eqn{B_1(a) \cap B_2}.
}
\description{
Compute the projection of a vector onto the intersection of the l1 ball
and the l2 ball.
}
\examples{
proj12(1:10, 21)
}
