% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj12orth.R
\name{proj12orth}
\alias{proj12orth}
\title{Compute the projection of a vector onto the intersection of the l1 ball,
the l2 ball and the space orthogonal to M.}
\usage{
proj12orth(x, a = 1, M, itermax = 5000, eps = 1e-16)
}
\arguments{
\item{x}{A vector of numerics}

\item{a}{The radius (>0) of the l1 ball}

\item{M}{A matrix of vectors}

\item{itermax}{The maximum number of iterations}

\item{eps}{Precision}
}
\value{
The projection of \eqn{x} onto \eqn{B_1(a) \cap B_2 \cap M^\perp}.
}
\description{
Compute the projection of a vector onto the intersection of the l1 ball,
the l2 ball and the space orthogonal to M.
}
\examples{
proj12orth(1:10, a=10, M=normalize(rnorm(10)))
}
