% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{projection}
\alias{projection}
\alias{projl1}
\alias{proj12}
\alias{proj12orth}
\alias{projgroup}
\alias{projorth}
\title{Compute the projection of a vector onto the l1 ball.}
\usage{
projl1(v, a)

proj12(x, a = 1)

proj12orth(x, a = 1, M, itermax = 5000, eps = 1e-16)

projgroup(v, g, a)

projorth(x, M)
}
\arguments{
\item{v}{A vector of numerics}

\item{a}{The radius (>0) of the l1 ball}

\item{x}{A vector of numerics}

\item{M}{A matrix of vectors}

\item{itermax}{The maximum number of iterations}

\item{eps}{Precision}

\item{g}{A factor describing the groups}
}
\value{
The projection of \eqn{v} onto the l1 ball of radius \eqn{a}.

The projection of \eqn{x} onto \eqn{B_1(a) \cap B_2}.

The projected vector and number of iterations.

The projection of \eqn{v} onto the group l1 ball of radius \eqn{a}.

The projection of \eqn{x} onto \eqn{V^\perp}.
}
\description{
Compute the projection of a vector onto the l1 ball.

Compute the projection onto the intersection of the l1 and l2 balls.

Compute the projection onto the intersection of l1, l2 and orthogonal space.

Compute the projection onto a group l1 ball.

Compute the projection onto the space orthogonal to a matrix.
}
\examples{
projl1(1:10, 21)
proj12(1:10, 21)
proj12orth(1:10, a = 10, M = normalize(rnorm(10)))
projgroup(1:10, rep(1:2, each = 5), 21)
projorth(1:10, normalize(rnorm(10)))
}
