% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projl1.R
\name{projl1}
\alias{projl1}
\title{Compute the projection of a vector onto the l1 ball.}
\usage{
projl1(v, a)
}
\arguments{
\item{v}{A vector of numerics}

\item{a}{The radius (>0) of the l1 ball}
}
\value{
The projection of \eqn{v} onto the l1 ball of radius \eqn{a}.
}
\description{
Compute the projection of a vector onto the l1 ball.
}
\examples{
projl1(1:10, 21)
}
