% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projorth.R
\name{projorth}
\alias{projorth}
\title{Compute the projection of a vector onto the space orthogonal to a given 
matrix.}
\usage{
projorth(x, M)
}
\arguments{
\item{x}{A vector of numerics}

\item{M}{A matrix of vectors}
}
\value{
The projection of \eqn{x} onto \eqn{V^\perp}.
}
\description{
Compute the projection of a vector onto the space orthogonal to a given 
matrix.
}
\examples{
projorth(1:10, normalize(rnorm(10)))
}
