% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{prox}
\alias{prox}
\alias{proxl1l2}
\title{Compute the group l1,2 proximal operator.}
\usage{
proxl1l2(u, g, a)
}
\arguments{
\item{u}{A vector of numerics}

\item{g}{A factor describing the groups}

\item{a}{The proximal threshold}
}
\value{
The group soft-thresholded vector.
}
\description{
Compute the group l1,2 proximal operator.
}
