% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_csvd.R
\name{theme_csvd}
\alias{theme_csvd}
\title{A ggplot2 theme that matches the defaults
of the factorial figures in PTCA4CATA}
\usage{
theme_csvd(mire = TRUE, x0 = 0, y0 = 0)
}
\arguments{
\item{mire}{(Default = \code{TRUE}) when \code{TRUE}
plot the factorial axis at values \code{x0} and \code{y0}.}

\item{x0}{(Default = 0) the X-axis origin. Relevant only
when \code{mire = TRUE}.}

\item{y0}{(Default = 0) the Y-axis origin.
Relevant only
when \code{mire = TRUE}.}
}
\description{
\code{theme_ptca}
A ggplot2 theme that matches the defaults
of the factorial figures in PTCA4CATA.
}
\examples{
\dontrun{
x = c(5, 8, 1, 70)
y = c(10, 12, 20, 50)
p <-  ggplot(as.data.frame(cbind(x,y)),
                 aes(x = x, y = y)) +
                    geom_point(size=2, shape=19) + theme_ptca()
}
}
\author{
Vincent Guillemot
}
